Revise pattern
